// ui/src/generated/VersionCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProjectId } from "./ProjectId";

export type VersionCommand = { "type": "List" } | { "type": "Create", name: string | null, } | { "type": "Restore", version: string, } | { "type": "Delete", version: string, } | { "type": "Rename", version: string, name: string | null, } | { "type": "Compare", version: string, against: string | null, } | { "type": "ListRecoverable" } | { "type": "Recover", project: ProjectId, } | { "type": "DiscardRecovery", project: ProjectId, } | { "type": "SessionStatus" };